requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
defusedxml==0.7.0rc2
Django==3.0.7
django-ckeditor==6.0.0
django-classy-tags==2.0.0
django-cookie-law==2.0.5
django-crispy-forms==1.9.2
django-js-asset==1.2.2
djangorestframework==3.12.2
idna==2.10
mysqlclient==2.0.3
oauthlib==3.1.0
Pillow==8.1.0
pycparser==2.20
PyJWT==2.0.1
pyserial==3.5
pytest-runner==5.3.1
python-gsmmodem-new==0.12
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.3.1
urllib3==1.25.11
